[#93234240] hooking in cc client for org creation
we are now injecting the cloud controlle client, which
will allow us to generate requests to the cc api
that are decorated with the proper auth token.

adding a file to keep track of user provided services created
for this app, in advance of pipeline management.

changes reflected across initroute, main and controller